requirements.txt
python-dotenv
openai
langchain
chromadb
tiktoken